Added and later disabled Russian roulette.

+++ Raytracer/StatsReporter.cpp
@@ -2,8 +2,8 @@
 
 int StatsReporter::numBVHNodes = 0;
 int StatsReporter::numBVHLeafNodes = 0;
-int StatsReporter::numRays = 0;
-int StatsReporter::numRayBoxIntersections = 0;
-int StatsReporter::numRayTriangleIntersections = 0;
+unsigned long StatsReporter::numRays = 0;
+unsigned long StatsReporter::numRayBoxIntersections = 0;
+unsigned long StatsReporter::numRayTriangleIntersections = 0;
 double StatsReporter::BVHBuildTime = 0.0;
 double StatsReporter::renderTime = 0.0;
